+++ plugins/providers/virtualbox/action/set_name.rb
@@ -12,6 +12,13 @@ module VagrantPlugins
         def call(env)
           name = env[:machine].provider_config.name
 
+          # If we already set the name before, then don't do anything
+          sentinel = env[:machine].data_dir.join("action_set_name")
+          if !name && sentinel.file?
+            @logger.info("Default name was already set before, not doing it again.")
+            return @app.call(env)
+          end
+
           # If no name was manually set, then use a default
           if !name
             prefix = "#{env[:root_path].basename.to_s}_#{env[:machine].name}"
@@ -30,6 +37,11 @@ module VagrantPlugins
             @logger.info("Setting the name of the VM: #{name}")
             env[:ui].info(I18n.t("vagrant.actions.vm.set_name.setting_name"))
             env[:machine].provider.driver.set_name(name)
+          end
+
+          # Create the sentinel
+          sentinel.open("w") do |f|
+            f.write(Time.now.to_i.to_s)
           end
 
           @app.call(env)